Fix detection of stale data (#63)
The function that checks is weather data is more than 2 hours old actually rejected data that was more than 2 hours older or younger than the present time of the raspberry. Unfortunately, raspberries that freshly boot can have a time that is not accurate. Assumption is that the raspberry will always have a time that is later than the real time.
The new approach is that weather data that is more than 2 hours older than the current date is rejected, but if the weatherdata is newer than the current time, it will be kept.